Broke everything out into individual files to make it more structured.
