+++ api/src/models/budget.rs
@@ -3,24 +3,16 @@ use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::collections::HashMap;
 use std::fmt::{Display, Formatter};
-use std::hash::{Hash, Hasher};
 use std::marker::PhantomData;
 use std::str::FromStr;
 use serde::de::{MapAccess, Visitor};
 use serde::ser::SerializeMap;
 use uuid::Uuid;
-
-#[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
-pub enum MonthBeginsOn {
-    PreviousMonth(u32),
-    CurrentMonth(u32),
-}
-
-impl Default for MonthBeginsOn {
-    fn default() -> Self {
-        MonthBeginsOn::PreviousMonth(25)
-    }
-}
+use crate::models::budget_category::BudgetCategory;
+use crate::models::budget_item::BudgetItem;
+use crate::models::budget_transaction::BudgetTransaction;
+use crate::models::budget_transaction_type::BudgetTransactionType;
+use crate::models::month_begins_on::MonthBeginsOn;
 
 #[derive(Debug, Clone, Serialize, Deserialize, Eq, Default, Model)]
 pub struct Budget {
@@ -252,79 +244,6 @@ impl Budget {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default, Model)]
-pub struct BankTransaction {
-    pub id: Uuid,
-    pub text: String,
-    pub amount: f32,
-    pub budget_item: Uuid,
-    pub bank_date: chrono::NaiveDate,
-    pub created_at: chrono::NaiveDateTime,
-    pub updated_at: chrono::NaiveDateTime,
-    pub created_by: Uuid,
-}
-
-impl Eq for BankTransaction {}
-
-impl BankTransaction {
-    pub fn new_from_user(
-        text: &str,
-        amount: f32,
-        budget_item: Uuid,
-        bank_date: chrono::NaiveDate,
-        created_by: Uuid,
-    ) -> Self {
-        Self {
-            id: Uuid::new_v4(),
-            text: text.to_string(),
-            amount,
-            budget_item,
-            bank_date,
-            created_at: chrono::Utc::now().naive_utc(),
-            updated_at: chrono::Utc::now().naive_utc(),
-            created_by,
-        }
-    }
-}
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Hash, Eq)]
-pub enum BudgetCategory {
-    Income(String),
-    Expense(String),
-    Savings(String),
-}
-
-impl Default for BudgetCategory {
-    fn default() -> Self {
-        BudgetCategory::Expense("Övrigt".to_string())
-    }
-}
-
-impl Display for BudgetCategory {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            BudgetCategory::Income(s) => write!(f, "Income({})", s),
-            BudgetCategory::Expense(s) => write!(f, "Expense({})", s),
-            BudgetCategory::Savings(s) => write!(f, "Savings({})", s),
-        }
-    }
-}
-
-
-impl FromStr for BudgetCategory {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        // Basic format: "VariantName(value)"
-        if let Some(rest) = s.strip_prefix("Income(").and_then(|s| s.strip_suffix(")")) {
-            return Ok(BudgetCategory::Income(rest.to_string()));
-        } else if let Some(rest) = s.strip_prefix("Expense(").and_then(|s| s.strip_suffix(")")) {
-            return Ok(BudgetCategory::Expense(rest.to_string()));
-        } else if let Some(rest) = s.strip_prefix("Savings(").and_then(|s| s.strip_suffix(")")) {
-            return Ok(BudgetCategory::Savings(rest.to_string()));
-        }
-        Err(format!("Unknown BudgetCategory format: {}", s))
-    }
-}
 
 fn deserialize_budget_items_with_string_keys<'de, D>(
     deserializer: D,
@@ -376,178 +295,3 @@ where
     map_ser.end()
 }
 
-#[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq, Default)]
-pub enum BudgetItemPeriodicity {
-    Once,
-    #[default]
-    Monthly,
-    Quarterly,
-    Yearly,
-}
-
-impl Display for BudgetItemPeriodicity {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            BudgetItemPeriodicity::Once => write!(f, "Once"),
-            BudgetItemPeriodicity::Monthly => write!(f, "Monthly"),
-            BudgetItemPeriodicity::Quarterly => write!(f, "Quarterly"),
-            BudgetItemPeriodicity::Yearly => write!(f, "Yearly"),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default, Model)]
-pub struct BudgetItem {
-    pub id: Uuid,
-    pub name: String,
-    pub budget_category: BudgetCategory,
-    pub incoming_transactions: HashMap<Uuid, BudgetTransaction>,
-    pub outgoing_transactions: HashMap<Uuid, BudgetTransaction>,
-    pub bank_transactions: HashMap<Uuid, BankTransaction>,
-    pub created_at: chrono::NaiveDateTime,
-    pub updated_at: chrono::NaiveDateTime,
-    pub created_by: Uuid,
-    pub budget_id: Uuid,
-}
-
-impl Hash for BudgetItem {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.id.hash(state);
-    }
-}
-
-impl BudgetItem {
-    pub fn new(
-        budget_id: Uuid,
-        name: &str,
-        budget_category: &BudgetCategory,
-        created_by: Uuid,
-    ) -> Self {
-        Self {
-            id: Uuid::new_v4(),
-            budget_id,
-            name: name.to_string(),
-            budget_category: budget_category.clone(),
-            created_at: chrono::Utc::now().naive_utc(),
-            updated_at: chrono::Utc::now().naive_utc(),
-            created_by,
-            ..Default::default()
-        }
-    }
-    
-    pub fn incoming_amount(&self) -> f32 {
-        self.incoming_transactions.values().map(|v| v.amount)
-            .sum::<f32>()
-    }
-    
-    pub fn outgoing_amount(&self) -> f32 {
-        self.outgoing_transactions.values().map(|v| v.amount)
-            .sum::<f32>()
-    }
-    
-    pub fn budgeted_amount(&self) -> f32 {
-        self.incoming_amount() - self.outgoing_amount()
-    }
-    
-    pub fn total_bank_amount(&self) -> f32 {
-        self.bank_transactions.values().map(|v| v.amount)
-            .sum::<f32>()
-    }
-
-    pub fn store_incoming_transaction(&mut self, budget_transaction: &BudgetTransaction) {
-        match self.incoming_transactions.entry(budget_transaction.id) {
-            Vacant(e) => {
-                e.insert(budget_transaction.clone());
-            }
-            Occupied(mut e) => {
-                e.insert(budget_transaction.clone());
-            }
-        }
-        self.touch();
-    }
-
-    pub fn store_outgoing_transaction(&mut self, budget_transaction: &BudgetTransaction) {
-        match self.outgoing_transactions.entry(budget_transaction.id) {
-            Vacant(e) => {
-                e.insert(budget_transaction.clone());
-            }
-            Occupied(mut e) => {
-                e.insert(budget_transaction.clone());
-            }
-        }
-        self.touch();
-    }
-
-    pub fn store_bank_transaction(&mut self, budget_transaction: &BankTransaction) {
-        match self.bank_transactions.entry(budget_transaction.id) {
-            Vacant(e) => {
-                e.insert(budget_transaction.clone());
-            }
-            Occupied(mut e) => {
-                e.insert(budget_transaction.clone());
-            }
-        }
-        self.touch();
-    }
-
-    pub fn touch(&mut self) {
-        self.updated_at = chrono::Utc::now().naive_utc();
-    }
-}
-
-#[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq, Eq, Hash, Default)]
-pub enum BudgetTransactionType {
-    #[default]
-    StartValue,
-    Adjustment,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default, Model)]
-pub struct BudgetTransaction {
-    pub id: Uuid,
-    pub text: String,
-    pub transaction_type: BudgetTransactionType,
-    pub amount: f32,
-    pub from_budget_item: Option<Uuid>,
-    pub to_budget_item: Uuid,
-    pub created_at: chrono::NaiveDateTime,
-    pub updated_at: chrono::NaiveDateTime,
-    pub created_by: Uuid,
-}
-
-impl PartialEq for BudgetTransaction {
-    fn eq(&self, other: &Self) -> bool {
-        self.id == other.id
-    }
-}
-
-impl Eq for BudgetTransaction {}
-
-impl Hash for BudgetTransaction {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.id.hash(state);
-    }
-}
-
-impl BudgetTransaction {
-    pub fn new(
-        text: &str,
-        transaction_type: BudgetTransactionType,
-        amount: f32,
-        from_budget_item: Option<Uuid>,
-        to_budget_item: Uuid,
-        created_by: Uuid,
-    ) -> Self {
-        Self {
-            id: Uuid::new_v4(),
-            text: text.to_string(),
-            transaction_type,
-            amount,
-            to_budget_item,
-            from_budget_item,
-            created_at: chrono::Utc::now().naive_utc(),
-            updated_at: chrono::Utc::now().naive_utc(),
-            created_by,
-        }
-    }
-}
