+++ api/src/models/budget_transaction_type.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Copy, Clone, Serialize, Deserialize, PartialEq, Eq, Hash, Default)]
+pub enum BudgetTransactionType {
+    #[default]
+    StartValue,
+    Adjustment,
+}
