+++ api/src/models/mod.rs
@@ -1,2 +1,19 @@
-pub mod budget;
-pub mod user;
+mod budget;
+mod user;
+mod bank_transaction;
+mod budget_category;
+mod budget_item;
+mod budget_transaction;
+mod budget_item_periodicity;
+mod budget_transaction_type;
+mod month_begins_on;
+
+pub use budget::Budget;
+pub use user::User;
+pub use bank_transaction::BankTransaction;
+pub use budget_category::BudgetCategory;
+pub use budget_item::BudgetItem;
+pub use budget_transaction::BudgetTransaction;
+pub use budget_item_periodicity::BudgetItemPeriodicity;
+pub use budget_transaction_type::BudgetTransactionType;
+pub use month_begins_on::MonthBeginsOn;
