+++ api/tests/budget_item_tests.rs
@@ -1,9 +1,8 @@
 #[cfg(test)]
 mod budget_item_tests {
-    use super::*;
-    use chrono::{NaiveDate, Utc};
+    use chrono::NaiveDate;
     use uuid::Uuid;
-    use api::models::budget::{BankTransaction, BudgetCategory, BudgetItem, BudgetTransaction, BudgetTransactionType};
+    use api::models::*;
 
     /// Helper function to create a test BudgetItem
     fn create_test_budget_item() -> BudgetItem {
